+++ CHANGELOG.md
@@ -12,10 +12,11 @@ being manually edited by hand.
     - Editor background:  #002540 -> #002138
     - Side Bar background: #002138 -> #001E33
     - Status Bar background: #001E33 -> #002540
+    - Current line: #003154 -> #002C4C
 - Booleans and numbers are now *italic*.
 - Properties now use the tag color. (Red in variants.)
 - Escape characters now use the function color. (Blue in variants.)
-- JS Constants and Objects now use the regex color. (Green in variants.)
+- JS Constants and Objects now use the component color. (Green in variants.)
 - Tabs now have a bottom border when active.
 - Reworked selection highlighting.